Move APIs to Parameter.cpp
